corrected indentation to 2 spaces / one tab, separated in diff test files and created additional stubs
