+++ models/user.js
@@ -1,32 +1,36 @@
 Chatter.User = ChatterUser = Astro.Class({
-    name: "ChatterUser",
-    collection: new Mongo.Collection("chatteruser"),
+  name: "ChatterUser",
+  collection: new Mongo.Collection("chatteruser"),
 
-    fields: {
-        userType: {
-            type: "string"
-        },
-        userId: {
-            type: "string"
-        },
-        nickname: {
-            type: "string"
-        },
-        avatar: {
-            type: "string",
-            default: function() {
-              return "http://localhost:3000/packages/jorgeer_chatter-semantic/public/images/avatar.jpg";
-            }
-        }
+  fields: {
+    userType: {
+      type: "string"
     },
-
-    events: {
-
+    userId: {
+      type: "string"
     },
-
-    methods: {
-
+    nickname: {
+      type: "string"
     },
+    avatar: {
+      type: "string",
+      default: function() {
+        return "http://localhost:3000/packages/jorgeer_chatter-semantic/public/images/avatar.jpg";
+      }
+    }
+  },
 
-    behaviors: ["timestamp"]
+  validators: {
+
+  },
+
+  events: {
+
+  },
+
+  methods: {
+
+  },
+
+  behaviors: ["timestamp"]
 });
